requirements.txt
flask
opencv-python-headless
face_recognition
pillow
numpy